表情界面行列不同化

+++ 表情键盘/Classes/Compose(发微博)/View/HMComposePhotosView.m
@@ -23,7 +23,7 @@
 {
     [super layoutSubviews];
     
-    int count = self.subviews.count;
+    int count = (int)self.subviews.count;
     // 一行的最大列数
     int maxColsPerRow = 4;
